feat(color/hdrimage/error): tone mapping implementation
* `s/usize/u32` inside `hdrimage.rs` and `error.rs`
* when `usize` needed `as usize` is the right workflow
* refactor unit tests (new `assert!(matches!())` method)
* unit test for tone mapping

+++ src/error.rs
@@ -3,7 +3,7 @@ use thiserror::Error;
 #[derive(Error, Debug)]
 pub enum HdrImageErr {
     #[error("invalid indexes: {0:?}, expected hdr image shape: {1:?}")]
-    OutOfBounds((usize, usize), (usize, usize)),
+    OutOfBounds((u32, u32), (u32, u32)),
     #[error("invalid pfm file format: {0}")]
     InvalidPfmFileFormat(String),
     #[error("impossible to read binary data from the file")]
